refactorizar funcionalidad a funcion get users
