+++ routes/users.js
@@ -77,26 +77,7 @@ module.exports = (app, next) => {
    * @code {401} si no hay cabecera de autenticación
    * @code {403} si no es ni admin
    */
-  app.get('/users', requireAdmin, (req, resp) => {
-    console.log('access-token: ', req.headers['access-token']);
-    // findAll metodo que recorre filas y retorna los arreglos
-    schemeTablaUser.findAll()
-      .then((data) => {
-        const newFormat = data.map((user) => {
-          const objectData = {
-            id: user.dataValues.id,
-            email: user.dataValues.email,
-            password: user.dataValues.password,
-            roles: {
-              admin: user.dataValues.roles,
-            },
-          };
-          return objectData;
-        });
-        resp.status(200).json(newFormat);
-      })
-      .catch((error) => { resp.status(500).json({ error: error.message }); });
-  });
+  app.get('/users', requireAdmin, getUsers);
 
   /**
    * @name GET /users/:uid
